Stepwise update on GraphAttributes: xTickmarks
Motivation
Final aim: I would like to add multiple curves in one lineChart.
This option is currently not available and several steps are required to allow this option.
In my opinion, the easiest way is to provide a list of data (as List (List Point)) instead
of data (as List Point). In order to distinguish the curves, they should have individual
options (e.g. color, strokeWidth), however ticks and scale should be identical for all curves.
Therefore, the GraphAttributes should be restructured: xTickmarks, yTickmarks, and scale
should be direct entries of the type alias GraphAttributes, and not part of the options.

Modification
- xTickmarks as direkt entry in record alias GraphAttributes, not part of type Option
- function xTickmarks options replaced by ga.xTickmarks (function obsolete)
- function boundingBox previously using List Option, now additionally using
  GraphAttributes (xTickmarks options replaced by ga.xTickmarks)
- function xTickmarks and xTickmarks: obsolete

Result
A part of the restructuring work is accomplished. xTickmarks are now direct entries
of GraphAttributes, and not part of the options.