Stepwise update on GraphAttributes: yTickmarks
Motivation
Final aim: I would like to add multiple curves in one lineChart.
This option is currently not available and several steps are required to allow this option.
In my opinion, the easiest way is to provide a list of data (as List (List Point)) instead
of data (as List Point). In order to distinguish the curves, they should have individual
options (e.g. color, strokeWidth), however ticks and scale should be identical for all curves.
Therefore, the GraphAttributes should be restructured: xTickmarks, yTickmarks, and scale
should be direct entries of the type alias GraphAttributes, and not part of the options.

Modification
- yTickmarks as direkt entry in record alias GraphAttributes, not part of type Option
- function yTickmarks ga.options replaced by ga.yTickmarks (function obsolete)
- function boundingBox previously using List Option and GraphAttributes,
  now only using GraphAttributes (yTickmarks options replaced by ga.yTickmarks)

Result
A part of the restructuring work is accomplished. yTickmarks are now direct entries
of GraphAttributes, and not part of the options.