Plotting several datasets in one graph
Motivation
The original version of elm-graph only allows one dataset to be plotted
as a linegraph. However, for my project, I need several datasets to be
displayed.

Modification
I created a variation of lineChartWithDataWindow that has the following
input:
- List (GraphAttributes) (instead of GraphAttributes)
- List (List (Float, Float)) (instead of List (Float, Float))
The GraphAttributes should be identical, except for the color to guarantee
proper functionality.
The function evaluates the data to facilitate the calculation of a proper
data window.
The function getDataWindow must be exposed from SimpleGraph.elm

Result
An additional function is provided to plot a list of datasets.
Some extra steps might be required to make the plot look good.

Solves https://github.com/jxxcarlson/elm-graph/issues/4