Restructuring
Motivation
This fork is supposed to support a specific type of graphic for
https://gitlab.desy.de/melanie.nentwich/oscars_test_app.
Aim is to provide the necessary functions in a minimalist way.

Modification
The newly developped function lineChartsWithDataWindow is now part of
the module SimpleGraph.elm and not of the example Main.elm.

Result
This fork of elm-graph is ready to be used in the target project.