Handling segments with zero-length
Motivation
Currently, lines and segments with zero-length (identical start and end)
are not visible. However, for the Beamline Finder app, it is important
that to mark those positions.

Modification
The function segmentToSVG now checks if start and end point are different
and then performs the regular commands. Else a circle is printed at the
corresponding position.

Result
Segments with zero-length are now visible.